Looked briefly at the reset and couldn't immediately find an issue. Fixed a couple of the uppers, and followed the logic. I did find that typing "y" exits instantly, and typing "Y" puts it in a really strange loop where it keeps all the old information and just asks for a guess over and over again.
I did fix all the answers coming out as wrong no matter what you type, as the random word that is created is all lower case and I was checking for upper(). That being said, now every time the word updates it's not remembering the previously correct guesses. I imagine this is a quick fix, but we will see.