Created a helper function for play_game() that seemed to fix most of the wonky recursive issues. That being said, there's still some clean up to do, as things aren't quite working as intended when the game is finished and the word is correct. It's skipping "Incredible! You win!" and there are no reset prompts. I will keep looking that over.
I also decided I'm likely going to move the reset_game function to main.py instead of having it in its own file. I think it's creating too many weird issues that started with circular imports and now are leading to some strange behavior. Moving it to main.py makes more sense, anyway, and then it can be treated as more of a helper function for main().