Deleted the reset file, since I didn't need it anymore.
When testing, I ran into a problem where answering a few incorrect answers before getting the corect word would just end the program without congratulating the user or asking for a reset. play_hangman.play_game had an issue in the "if guess_bool is False" portion, where it would return after calling keep_playing and would exit. I fixed this by adding in a conditional that if the word is finished/the user solved it, then it would return.